+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.3
+
+- Allow `LiquidStream::Streams` to have a default source
+- For `LiquidStream::Streams`, if no source nor default source is defined, then source is an empty array
+
 # 0.0.2
 
 Fixes issue with Utils being in the wrong place and broke loading